ros2_ws/src/teleop_pipeline/requirements.txt
# Python dependencies for teleop_pipeline package
numpy>=1.20.0
mujoco>=3.0.0
mink-mujoco>=0.1.0
scipy>=1.7.0
robot-descriptions>=1.8.0

